refactor: consolidate endpoints to one main.py, changed Makefile and Dockerfile accordingly
