+++ prediction_service/Makefile
@@ -1,11 +1,8 @@
 run-dev:
-	uvicorn main:app --host 0.0.0.0 --port 8081
-
-run-dev-auth:
-	uvicorn main_auth:app --host 0.0.0.0 --port 8082
+	uvicorn main:app --host 0.0.0.0 --port 8082
 
 build: 
-	docker build --no-cache -t prediction-service .
+	docker build --no-cache -t telco-churn-prediction-service .
 
 run-live: build
-	docker run -p 8082:8082 prediction-service
+	docker run -p 8082:8082 telco-churn-prediction-service
