+++ prediction_service/api_keys.env
@@ -1,0 +1,1 @@
+API_KEY = "doImMAExEOi40hrNG5XKBmyARRRWTue_8qTEk4XKnas, 7IfumrmpJ6oOrTK4dVDArkNFpRScVkZEF66Qb2Ke6ic"
